+++ ProjectDirectX/GraphicsHandler.h
@@ -43,7 +43,7 @@ public:
 
 	bool Initialize(int screenWidth, int screenHeight, HWND hwnd);
 	void ShutDown();
-	bool Frame(float fps, float frameTime, InputHandler* inputObj);
+	bool Frame(int fps, float frameTime, InputHandler* inputObj);
 	bool Render();
 
 private:
